[JENKINS-61438] IllegalArgumentException for ChoiceParameter on bindJSON saving job configuration
Fix potential NPE
